fitlog-backend-56-adressed PR comments
